complete all game loops mechanics?
- game over scene
- ghost angry basic feedback
- patience/mood slider
- wrong order mechanics

+++ FolkloreGameJam/Assets/Scenes/Scripts/ScriptableObject/Menu.cs
@@ -14,5 +14,5 @@ public class Menu : ScriptableObject
     [Header("Menu Properties")]
     public FoodType FoodType;
     public Sprite Sprite;
-    public float Score;
+    public int Score;
 }
